+++ src/Sigma/Doppma.cs
@@ -73,7 +73,7 @@ public class Doppma : BaseSigma {
 				lastAttackFrame = Global.level.frameCount;
 			}
 		}
-		framesSinceLastAttack = Global.level.frameCount - lastAttackFrame;
+		framesSinceLastAttack = (int)(Global.level.frameCount - lastAttackFrame);
 		bool lenientAttackPressed = (attackPressed || framesSinceLastAttack < 5);
 
 		// Shoot button attacks.
